firedns.h: make v6 connection member flag a char, and always used
otherwise the struct size will mismatch when compiled with or without
ipv6 support, and thus messing up our state struct when lib and prog
were compiled with different settings.

since it was only used as flag, we use signed char instead and get it
for free in the alignment overhead.